WIP - Story Guide
Based on discussion Kevin and I had about having our own story guide, points above are lifted from decks that Taylor and I assembled a while back to form a guide on story practices:
* Storywriting Best Practices
* Stoywriting/JIRA audit

I've added some more content and moved sections around in an attempt to better package this as a guide. Hopefully, this has not lost the spirit of the original decks' value.

Feedback requested and welcome! Thanks

@taykcrane @bechols @krista-thomas @yasicheh @kevgo @matthandlersux @danstorms